+++ services/sso/src/routes/dashboard/usersRoutes.ts
@@ -5,10 +5,12 @@ import { pathRoot } from '../routes'
 import { dashboardListUsers } from '../../controllers/dashboard/users/list'
 import { parse, validate } from '../../middleware/validate'
 import { z } from '../../openapi/zod'
-import { dashboardUsersListQuerySchema } from '../../schemas'
 import { getMe } from '../../controllers/dashboard/users/getMe'
+import { dashboardUsersListQuerySchema } from '../../schemas'
+import { dashboardUserUpdateSchema } from '../../schemas/users/dashboardUserUpdateSchema'
 import { userIdSchema } from '../../schemas/users/userSchema'
 import { dashboardDeleteUser } from '../../controllers/dashboard/users/delete'
+import { updateUser } from '../../controllers/users/update'
 
 export const dashboardUsersRoutes = new Router()
 
@@ -25,6 +27,19 @@ dashboardUsersRoutes.get(
   dashboardListUsers
 )
 dashboardUsersRoutes.get('/me', authenticate, authorize('ADMIN'), getMe)
+
+dashboardUsersRoutes.patch(
+  '/:id',
+  authenticate,
+  authorize('ADMIN'),
+  validate(
+    z.object({
+      params: z.object({ id: userIdSchema }),
+      body: dashboardUserUpdateSchema,
+    })
+  ),
+  updateUser
+)
 
 dashboardUsersRoutes.delete(
   '/:id',